Refatoração métodos de leitura e nó CSV
