+++ src/controller/Lista.java
@@ -11,6 +11,7 @@ import javax.swing.JOptionPane;
 import entity.No;
 import entity.Inscrito;
 
+
 public class Lista {
 	
 	String arquivo = "ListaAlunos.csv";
@@ -63,28 +64,14 @@ public class Lista {
 		return buscaUltimo(aux.getProx());
 	}
 	
-	public String percorre() {
-		String lista = "Cadastrados\n\n";
-		No aux = inicio;
-		lista = concatena(aux, lista);
-		return lista;
-	}
-
-	public String concatena(No aux, String lista) {
-		if (aux != null) {
-			lista += "Nome: " + aux.getInscrito().getNome() + "\nCurriculo: " + aux.getInscrito().getCurriculo()
-					+ "\nCPF: " + aux.getInscrito().getCpf() + "\nOpcCurso: " + aux.getInscrito().getOpcCurso()
-					+ "\nE-mail: " + aux.getInscrito().getEmail() + "\nRG: " + aux.getInscrito().getRg()
-					+ "\nTelefone: " + aux.getInscrito().getTelefone() +"\nFaculdade: " + aux.getInscrito().getNomeFaculdade() +
-					"\nCurso :" + aux.getInscrito().getNomeCurso() + "\nMédia geral: " + aux.getInscrito().getMediaFaculdade() 
-					+ "\n\n";
-			System.out.println(lista);
-			return concatena(aux.getProx(), lista);
+	public void percorre() {
+		try {
+			inscritoController.lerArquivo(arquivo);
+		} catch (IOException e) {
+			e.printStackTrace();
 		}
-
-		return lista;
 	}
-
+		
 	
 	public String removeFinal() {
 		if (inicio.getProx() == null) {
@@ -174,119 +161,7 @@ public class Lista {
 		}
 	}
 	
-	public void percorreCsv(String nome_arquivo) throws IOException{
-		inicio = null;
-		File arq = new File(nome_arquivo);
-		String nome="", curriculo="", cpf="", curso="", email="", rg="", telefone="", nomeDaFaculdade="", nomeDoCurso="";
-		double mediaGeral = 0;
-		if (arq.exists() && arq.isFile()) {
-			FileInputStream fluxo = new FileInputStream(arq);
-			InputStreamReader leitor = new InputStreamReader(fluxo);
-			BufferedReader buffer = new BufferedReader(leitor);
-			String linha = buffer.readLine();
-			String[] dadosDoCandidato;
-			linha = buffer.readLine(); // pulando a primeira linha
-			while (linha != null) { // procurando EOF
-				dadosDoCandidato = linha.split(";");
-				nome = dadosDoCandidato[0];
-				curriculo = dadosDoCandidato[1];
-				cpf = dadosDoCandidato[2];
-				curso = dadosDoCandidato[3];
-				email = dadosDoCandidato[4];
-				rg = dadosDoCandidato[5];
-				telefone = dadosDoCandidato[6];
-				nomeDaFaculdade = dadosDoCandidato[7];
-				nomeDoCurso = dadosDoCandidato[8];
-				mediaGeral = Double.parseDouble(dadosDoCandidato[9]);
-				insereNo(nome, curriculo, cpf, curso, email, rg, telefone, nomeDaFaculdade, nomeDoCurso, mediaGeral);
-				linha = buffer.readLine();	
-			}
-			buffer.close();
-			leitor.close();
-			fluxo.close();
-		} else {
-			throw new IOException("Arquivo Inválido");
-		}
-	}
 
-	private void insereNo(String nome, String curriculo, String cpf, String curso, String email, String rg,
-			String telefone, String nomeDaFaculdade, String nomeDoCurso, double mediaGeral) {
-		No novo_no = new No();
-		Inscrito novo_inscrito = new Inscrito();
-		novo_inscrito.setNome(nome);
-		novo_inscrito.setCurriculo(curriculo);
-		novo_inscrito.setCpf(cpf);
-		novo_inscrito.setOpcCurso(curso);
-		novo_inscrito.setEmail(email);
-		novo_inscrito.setRg(rg);
-		novo_inscrito.setTelefone(telefone);
-		novo_inscrito.setNomeFaculdade(nomeDaFaculdade);
-		novo_inscrito.setNomeCurso(nomeDoCurso);
-		novo_inscrito.setMediaFaculdade(mediaGeral);
-		
-		novo_no.setInscrito(novo_inscrito);
-		
-		if (vazia()) {
-			
-			novo_no.setProx(null);
-			novo_no.setAnterior(null);
-			inicio = novo_no;
-
-		} else {
-
-			No aux;
-			aux = buscaUltimo(inicio);
-			aux.setProx(novo_no);
-			novo_no.setAnterior(aux);
-			novo_no.setProx(null);
-
-		}
-		
-	}
-	
-	public void lerArquivo(String nomeArquivo) throws IOException {
-		File arq = new File(nomeArquivo);
-		String nome="", curriculo="", cpf="", curso="", email="", rg="", telefone="", nomeDaFaculdade="", nomeDoCurso="", mediaGeral = "";
-		if (arq.exists() && arq.isFile()) {
-			FileInputStream fluxo = new FileInputStream(arq);
-			InputStreamReader leitor = new InputStreamReader(fluxo);
-			BufferedReader buffer = new BufferedReader(leitor);
-			String linha = buffer.readLine();
-			linha = buffer.readLine(); // pular a primeira linha
-			String[] dadosDoCandidato;
-			while (linha != null) { // procurando EOF
-				dadosDoCandidato = linha.split(";");
-				nome = dadosDoCandidato[0];
-				curriculo = dadosDoCandidato[1];
-				cpf = dadosDoCandidato[2];
-				curso = dadosDoCandidato[3];
-				email = dadosDoCandidato[4];
-				rg = dadosDoCandidato[5];
-				telefone = dadosDoCandidato[6];
-				nomeDaFaculdade = dadosDoCandidato[7];
-				nomeDoCurso = dadosDoCandidato[8];
-				mediaGeral = dadosDoCandidato[9];
-				String candidato = "Nome: "+nome+"\n"
-								+ "Curriculo: "+curriculo+"\n"
-								+ "CPF: "+cpf+"\n"
-								+ "Curso: "+curso+"\n"
-								+ "Email: "+email+"\n"
-								+ "RG: "+rg+"\n"
-								+ "Telefone: "+telefone+"\n"
-								+ "Faculdade: "+nomeDaFaculdade+"\n"
-								+ "Curso: "+nomeDoCurso+"\n"
-								+ "Média Geral: "+mediaGeral+"\n";
-				System.out.println(candidato);
-				System.out.println("___________________________________");
-				linha = buffer.readLine();
-			}
-			buffer.close();
-			leitor.close();
-			fluxo.close();
-		} else {
-			throw new IOException("Arquivo Inválido");
-		}
-	}
 	
 	public void ordenaArquivo(String nomeArquivo) {
 		// Em construção
